Refactor auth flow code and simplify window opening
